+++ init/init.php
@@ -98,6 +98,9 @@ $objectInstancier->set('ActesImapProperties',$actesImapProperties);
 $objectInstancier->set('actes_response_tmp_local_path',ACTES_RESPONSE_TMP_LOCAL_PATH);
 $objectInstancier->set('actes_response_error_path',ACTES_RESPONSE_ERROR_PATH);
 
+$objectInstancier->set('pades_valid_url',PADES_VALID_URL);
+$objectInstancier->set('rgs_validca_path',RGS_VALIDCA_PATH);
+
 if (php_sapi_name() != 'cli'){
     $objectInstancier->get('Logger')->setLogType(Logger::TYPE_MEMORY);
 }